+++ src/test/java/seedu/modsuni/storage/XmlSerializableUserTest.java
@@ -1,4 +1,4 @@
-package seedu.address.storage;
+package seedu.modsuni.storage;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -7,7 +7,7 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
-import seedu.address.commons.util.XmlUtil;
+import seedu.modsuni.commons.util.XmlUtil;
 
 public class XmlSerializableUserTest {
     private static final Path TEST_DATA_FOLDER = Paths.get("src", "test",
